requirements.txt
python-telegram-bot==21.3
python-dotenv==1.0.0
aiofiles==23.2.1
aiosqlite==0.20.0
telethon==1.36.0
cryptg==0.4.0
